Add external resource usage

+++ gcp/compute/disk.py
@@ -79,16 +79,28 @@ class Disk(GoogleCloudPlatform):
 @operation
 @utils.throw_cloudify_exceptions
 def create(image, name, **kwargs):
+    name = get_disk_name(name)
     gcp_config = utils.get_gcp_config()
-    name = name or utils.get_gcp_resource_name(ctx.instance.id)
     disk = Disk(gcp_config,
                 ctx.logger,
                 image=image,
                 name=name)
-    disk.create()
+    create_disk(disk)
     ctx.instance.runtime_properties[constants.NAME] = name
     ctx.instance.runtime_properties[constants.DISK] = \
         disk.disk_to_insert_instance_dict(name)
+
+
+def get_disk_name(name):
+    if utils.should_use_external_resource():
+        return utils.assure_resource_id_correct()
+    else:
+        return name or utils.get_gcp_resource_name(ctx.instance.id)
+
+
+def create_disk(disk):
+    if not utils.should_use_external_resource():
+        disk.create()
 
 
 @operation
@@ -99,9 +111,14 @@ def delete(**kwargs):
     disk = Disk(gcp_config,
                 ctx.logger,
                 name=name)
-    disk.delete()
+    delete_disk(disk)
     ctx.instance.runtime_properties.pop(constants.DISK, None)
     ctx.instance.runtime_properties.pop(constants.NAME, None)
+
+
+def delete_disk(disk):
+    if not utils.should_use_external_resource():
+        disk.delete()
 
 
 @operation
