+++ gcp/compute/firewall.py
@@ -118,15 +118,27 @@ class FirewallRule(GoogleCloudPlatform):
 def create(firewall_rule, **kwargs):
     gcp_config = utils.get_gcp_config()
     network_name = utils.get_gcp_resource_name(gcp_config['network'])
-    firewall_rule['name'] = utils.get_firewall_rule_name(network_name,
-                                                         firewall_rule)
+    set_firewall_rule_name(firewall_rule, network_name)
     firewall = FirewallRule(gcp_config,
                             ctx.logger,
                             firewall=firewall_rule,
                             network=network_name)
 
-    firewall.create()
+    create_firewall(firewall)
     ctx.instance.runtime_properties[constants.NAME] = firewall.name
+
+
+def set_firewall_rule_name(firewall_rule, network_name):
+    if utils.should_use_external_resource():
+        firewall_rule['name'] = utils.assure_resource_id_correct()
+    else:
+        firewall_rule['name'] = utils.get_firewall_rule_name(network_name,
+                                                             firewall_rule)
+
+
+def create_firewall(firewall):
+    if not utils.should_use_external_resource():
+        firewall.create()
 
 
 @operation
@@ -141,27 +153,41 @@ def delete(**kwargs):
                             ctx.logger,
                             firewall={'name': firewall_name},
                             network=network_name)
-    firewall.delete()
+    delete_firewall(firewall)
     ctx.instance.runtime_properties.pop(constants.NAME, None)
+
+
+def delete_firewall(firewall):
+    if not utils.should_use_external_resource():
+        firewall.delete()
 
 
 @operation
 @utils.throw_cloudify_exceptions
 def create_security_group(rules, **kwargs):
     gcp_config = utils.get_gcp_config()
-    firewall = utils.create_firewall_structure_from_rules(
+    firewall_structure = create_firewall_structure_from_rules(
         gcp_config['network'],
         rules)
     ctx.instance.runtime_properties[constants.TARGET_TAGS] = \
-        firewall[constants.TARGET_TAGS]
+        firewall_structure[constants.TARGET_TAGS]
     ctx.instance.runtime_properties[constants.SOURCE_TAGS] = \
-        firewall[constants.SOURCE_TAGS]
+        firewall_structure[constants.SOURCE_TAGS]
     firewall = FirewallRule(gcp_config,
                             ctx.logger,
-                            firewall,
+                            firewall_structure,
                             gcp_config['network'])
-    firewall.create()
     ctx.instance.runtime_properties[constants.NAME] = firewall.name
+    create_firewall(firewall)
+
+
+def create_firewall_structure_from_rules(network, rules):
+    firewall_structure = utils.create_firewall_structure_from_rules(
+        network,
+        rules)
+    if utils.should_use_external_resource():
+        firewall_structure['name'] = utils.assure_resource_id_correct()
+    return firewall_structure
 
 
 @operation
