+++ gcp/compute/instance.py
@@ -290,27 +290,40 @@ class Instance(GoogleCloudPlatform):
 @operation
 @utils.throw_cloudify_exceptions
 def create(instance_type, image_id, properties, name, **kwargs):
+    name = get_instance_name(name)
     gcp_config = utils.get_gcp_config()
     gcp_config['network'] = utils.get_gcp_resource_name(gcp_config['network'])
     script = properties.get('startup_script')
     if script:
         script = ctx.download_resource(script)
-    instance_name = name or ctx.instance.id
+    #name = name or utils.get_gcp_resource_name(ctx.instance.id)
     instance = Instance(gcp_config,
                         ctx.logger,
-                        name=instance_name,
+                        name=name,
                         image=image_id,
                         machine_type=instance_type,
                         external_ip=properties.get('externalIP', False),
                         startup_script=script)
+    ctx.instance.runtime_properties[constants.NAME] = instance.name
     if ctx.node.properties['install_agent']:
         add_to_security_groups(instance)
     disk = ctx.instance.runtime_properties.get(constants.DISK)
     if disk:
         instance.disks = [disk]
-    instance.create()
-    ctx.instance.runtime_properties[constants.NAME] = instance.name
+    create_instance(instance)
     set_ip(instance)
+
+
+def get_instance_name(name):
+    if utils.should_use_external_resource():
+        return utils.assure_resource_id_correct()
+    else:
+        return name or utils.get_gcp_resource_name(ctx.instance.id)
+
+
+def create_instance(instance):
+    if not utils.should_use_external_resource():
+        instance.create()
 
 
 @operation
@@ -323,9 +336,15 @@ def delete(**kwargs):
     instance = Instance(gcp_config,
                         ctx.logger,
                         name=name)
-    instance.delete()
+    delete_instance(instance)
+
     ctx.instance.runtime_properties.pop(constants.DISK, None)
     ctx.instance.runtime_properties.pop(constants.NAME, None)
+
+
+def delete_instance(instance):
+    if not utils.should_use_external_resource():
+        instance.delete()
 
 
 @operation
